requirements.txt
mss
keyboard
Pillow
psutil
google-genai
python-dotenv 
google-api-python-client